change id type to long

+++ comm/Node.hpp
@@ -46,7 +46,7 @@ template<typename T> class Node{
     inline void _insert(T data_in){ dataArr.push_back(data_in); }
     std::vector<NodeIndex> children;
     public:
-    Node(int id_in):id(id_in){ }
+    Node(long id_in):id(id_in){ }
 
     Node(T data_in, int id_in):id(id_in),parent(-1){ dataArr.push_back(data_in); }
